Added solution - LeetHub

+++ Height of Binary Tree - GFG/height-of-binary-tree.java
@@ -125,6 +125,8 @@ class Solution {
     {
         // code here 
         if(node==null)return 0;
-        return Math.max(height(node.left),height(node.right))+1;
+        return (int)Math.max(height(node.left),height(node.right))+1;
+        
     }
+    
 }
